changes implemented, 1 remark added as unresolved
